GPU kMeans - type fix and rename script.

+++ pero_pretraining/scripts/fit_kmeans_gpu.py
@@ -34,7 +34,7 @@ def KMeans(x, K=10, Niter=10, c=None, verbose=True):
         change = (last_assignment != cl).sum().item() if last_assignment is not None else 0
         last_assignment = cl
         # compute min, median, mean, and max number of points per cluster
-        cluter_sizes = torch.bincount(cl)
+        cluter_sizes = torch.bincount(cl).type_as(c)
         print(i, change, D_ij.min(dim=1).mean().item(),
               f"min: {cluter_sizes.min().item()}, median: {cluter_sizes.median().item()}, "
               f"mean: {cluter_sizes.mean().item()}, max: {cluter_sizes.max().item()}")
